# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/paper-wallet-core-service/users": {
            "get": {
                "description": "Retrieve a list of all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-controller"
                ],
                "summary": "Get all users",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iOS",
                        "description": "Channel identifier",
                        "name": "X-Channel-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "RequestId",
                        "description": "Unique request identifier",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "gateway",
                        "description": "Service identifier",
                        "name": "X-Service-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username associated with the request",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "id",
                        "description": "Language preference for the response",
                        "name": "Accept-Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "User-Agent: Mozilla/5.0 (iPhone; CPU iPhone OS 14_4_2 like Mac OS X",
                        "description": "User agent identifier",
                        "name": "User-Agent",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Basic dXNlcm5hbWU6cGFzc3dvcmQ=",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1.2.3",
                        "description": "Application version",
                        "name": "X-App-Version",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.User"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update the details of an existing user based on the provided user ID and data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-controller"
                ],
                "summary": "Update an existing user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iOS",
                        "description": "Channel identifier",
                        "name": "X-Channel-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "RequestId",
                        "description": "Unique request identifier",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "gateway",
                        "description": "Service identifier",
                        "name": "X-Service-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username associated with the request",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "id",
                        "description": "Language preference for the response",
                        "name": "Accept-Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "User-Agent: Mozilla/5.0 (iPhone; CPU iPhone OS 14_4_2 like Mac OS X",
                        "description": "User agent identifier",
                        "name": "User-Agent",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Basic dXNlcm5hbWU6cGFzc3dvcmQ=",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1.2.3",
                        "description": "Application version",
                        "name": "X-App-Version",
                        "in": "header"
                    },
                    {
                        "description": "User Data for Update",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateUserRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-controller"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iOS",
                        "description": "Channel identifier",
                        "name": "X-Channel-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "RequestId",
                        "description": "Unique request identifier",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "gateway",
                        "description": "Service identifier",
                        "name": "X-Service-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username associated with the request",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "id",
                        "description": "Language preference for the response",
                        "name": "Accept-Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "User-Agent: Mozilla/5.0 (iPhone; CPU iPhone OS 14_4_2 like Mac OS X",
                        "description": "User agent identifier",
                        "name": "User-Agent",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Basic dXNlcm5hbWU6cGFzc3dvcmQ=",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1.2.3",
                        "description": "Application version",
                        "name": "X-App-Version",
                        "in": "header"
                    },
                    {
                        "description": "User Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserResponseDto"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a user by their ID, either hard or soft delete based on the provided flag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-controller"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iOS",
                        "description": "Channel identifier",
                        "name": "X-Channel-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "RequestId",
                        "description": "Unique request identifier",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "gateway",
                        "description": "Service identifier",
                        "name": "X-Service-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username associated with the request",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "id",
                        "description": "Language preference for the response",
                        "name": "Accept-Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "User-Agent: Mozilla/5.0 (iPhone; CPU iPhone OS 14_4_2 like Mac OS X",
                        "description": "User agent identifier",
                        "name": "User-Agent",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Basic dXNlcm5hbWU6cGFzc3dvcmQ=",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1.2.3",
                        "description": "Application version",
                        "name": "X-App-Version",
                        "in": "header"
                    },
                    {
                        "description": "Delete User Request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.DeleteUserRequestDto"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User successfully deleted"
                    }
                }
            }
        },
        "/paper-wallet-core-service/users/{id}": {
            "get": {
                "description": "Get details of a user by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-controller"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iOS",
                        "description": "Channel identifier",
                        "name": "X-Channel-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "RequestId",
                        "description": "Unique request identifier",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "gateway",
                        "description": "Service identifier",
                        "name": "X-Service-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username associated with the request",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "id",
                        "description": "Language preference for the response",
                        "name": "Accept-Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "User-Agent: Mozilla/5.0 (iPhone; CPU iPhone OS 14_4_2 like Mac OS X",
                        "description": "User agent identifier",
                        "name": "User-Agent",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Basic dXNlcm5hbWU6cGFzc3dvcmQ=",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1.2.3",
                        "description": "Application version",
                        "name": "X-App-Version",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    }
                }
            }
        },
        "/paper-wallet-core-service/wallet/withdraw": {
            "post": {
                "description": "Withdraw funds from the user's wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet-controller"
                ],
                "summary": "Withdraw funds from wallet",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iOS",
                        "description": "Channel identifier",
                        "name": "X-Channel-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "RequestId",
                        "description": "Unique request identifier",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "gateway",
                        "description": "Service identifier",
                        "name": "X-Service-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username associated with the request",
                        "name": "X-Username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "id",
                        "description": "Language preference for the response",
                        "name": "Accept-Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "User-Agent: Mozilla/5.0 (iPhone; CPU iPhone OS 14_4_2 like Mac OS X",
                        "description": "User agent identifier",
                        "name": "User-Agent",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Basic dXNlcm5hbWU6cGFzc3dvcmQ=",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1.2.3",
                        "description": "Application version",
                        "name": "X-App-Version",
                        "in": "header"
                    },
                    {
                        "description": "Withdraw Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/withdraw.WithdrawRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Withdrawal Successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.User": {
            "type": "object",
            "required": [
                "balance",
                "currency",
                "name",
                "scale"
            ],
            "properties": {
                "balance": {
                    "type": "number"
                },
                "createdBy": {
                    "type": "string"
                },
                "createdDate": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scale": {
                    "type": "integer"
                },
                "updatedBy": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "user.CreateUserRequestDto": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/user.UserDto"
                }
            }
        },
        "user.CreateUserResponseDto": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/user.UserDto"
                }
            }
        },
        "user.DeleteUserRequestDto": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "isHardDelete": {
                    "type": "boolean"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "user.UpdateUserRequestDto": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/user.UserDto"
                }
            }
        },
        "user.UserDto": {
            "type": "object",
            "required": [
                "balance",
                "currency",
                "name",
                "scale"
            ],
            "properties": {
                "balance": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "scale": {
                    "type": "integer"
                }
            }
        },
        "withdraw.WithdrawRequestDto": {
            "type": "object",
            "required": [
                "amount",
                "userId"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}